refactored editTask to include functionality of newTask
these two pages had similar functionality so i just combined the pages and added some conditionals to determine which page was which. not sure if this is proper practice, but i removed a lot of duplicate code. at the same time i added a lot of conditionals tho which may make the readability harder...